feat(portal): create common fetch doc controller
